# api/swagger.yaml
---
swagger: "2.0"
info:
  description: "The Burpee Counter API"
  version: "1.0.0"
  title: "Burpee Counter"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "burpee-counter@googlegroups.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3001"
basePath: "/v1"
tags:
- name: "burpees"
  description: "Burpee Counter"
  externalDocs:
    description: "Find out more"
    url: "http://swagger.io"
schemes:
- "http"
paths:
  /burpees:
    get:
      tags:
      - "burpees"
      summary: "Get all burpees"
      description: ""
      operationId: "totalBurpees"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Burpee"
        "405":
          description: "Invalid input"
      security:
      - burpees_auth:
        - "read:burpees"
  /burpees/{userId}:
    get:
      tags:
      - "burpees"
      summary: "Find pet by ID"
      description: "Returns a single pet"
      operationId: "getPetById"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of pet to return"
        required: true
        type: "integer"
        format: "int64"
        x-exportParamName: "UserId"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Burpee"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Pet not found"
      security:
      - api_key: []
    post:
      tags:
      - "burpees"
      summary: "Updates a pet in the store with form data"
      description: ""
      operationId: "updatePetWithForm"
      consumes:
      - "application/x-www-form-urlencoded"
      produces:
      - "application/json"
      parameters:
      - name: "userId"
        in: "path"
        description: "ID of pet that needs to be updated"
        required: true
        type: "integer"
        format: "int64"
        x-exportParamName: "UserId"
      - name: "name"
        in: "formData"
        description: "Updated name of the pet"
        required: false
        type: "string"
        x-exportParamName: "Name"
        x-optionalDataType: "String"
      - name: "status"
        in: "formData"
        description: "Updated status of the pet"
        required: false
        type: "string"
        x-exportParamName: "Status"
        x-optionalDataType: "String"
      responses:
        "405":
          description: "Invalid input"
      security:
      - burpees_auth:
        - "write:pets"
        - "read:pets"
securityDefinitions:
  burpees_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
      read:burpees: "read your burpees"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
  Burpee:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      count:
        type: "integer"
        format: "int64"
        example: 2
    example:
      count: 2
      id: 0
externalDocs:
  description: "Find out more about Swagger"
  url: "http://swagger.io"