# go/model_burpee.go
/*
 * Burpee Counter
 *
 * The Burpee Counter API
 *
 * API version: 1.0.0
 * Contact: burpee-counter@googlegroups.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Burpee struct {

	Id int64 `json:"id,omitempty"`

	Count int64 `json:"count,omitempty"`
}